Renamed the download helper
